bug: move app description endpoints (#278)
move app description endpoint from app to appchange controller

Refs: CPLP-2345
