+++ src/marketplace/Apps.Service/BusinessLogic/AppChangeBusinessLogic.cs
@@ -27,6 +27,7 @@ using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Repositories;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
 using Org.Eclipse.TractusX.Portal.Backend.Provisioning.Library;
+using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
 using System.Text.Json;
 
 namespace Org.Eclipse.TractusX.Portal.Backend.Apps.Service.BusinessLogic;
@@ -97,5 +98,50 @@ public class AppChangeBusinessLogic : IAppChangeBusinessLogic
         await _notificationService.CreateNotifications(_settings.ActiveAppCompanyAdminRoles, result.CompanyUserId, content, result.ProviderCompanyId.Value).ConfigureAwait(false);
         await _portalRepositories.SaveAsync().ConfigureAwait(false);
         return roleData;
+    }
+
+    /// <inheritdoc />
+    public async Task<IEnumerable<LocalizedDescription>> GetAppUpdateDescriptionByIdAsync(Guid appId, string iamUserId)
+    {        
+        var offerRepository = _portalRepositories.GetInstance<IOfferRepository>();
+        return await ValidateAndGetAppDescription(appId, iamUserId, offerRepository);        
+    }
+    
+    /// <inheritdoc />
+    public async Task CreateOrUpdateAppDescriptionByIdAsync(Guid appId, string iamUserId, IEnumerable<LocalizedDescription> offerDescriptionDatas)
+    {
+        var offerRepository = _portalRepositories.GetInstance<IOfferRepository>();
+
+        offerRepository.CreateUpdateDeleteOfferDescriptions(appId,
+            await ValidateAndGetAppDescription(appId, iamUserId, offerRepository),
+            offerDescriptionDatas.Select(od => new ValueTuple<string,string, string>(od.LanguageCode,od.LongDescription,od.ShortDescription)));
+
+        await _portalRepositories.SaveAsync().ConfigureAwait(false);
+    }
+    
+    private static async Task<IEnumerable<LocalizedDescription>> ValidateAndGetAppDescription(Guid appId, string iamUserId, IOfferRepository offerRepository)
+    {
+        var result = await offerRepository.GetActiveOfferDescriptionDataByIdAsync(appId, OfferTypeId.APP, iamUserId).ConfigureAwait(false);
+        if (result == default)
+        {
+            throw new NotFoundException($"App {appId} does not exist.");
+        }
+
+        if (!result.IsStatusActive)
+        {
+            throw new ConflictException($"App {appId} is in InCorrect Status");
+        }
+
+        if (!result.IsProviderCompanyUser)
+        {
+            throw new ForbiddenException($"user {iamUserId} is not a member of the providercompany of App {appId}");
+        }
+
+        if (result.OfferDescriptionDatas == null)
+        {
+            throw new UnexpectedConditionException("offerDescriptionDatas should never be null here");
+        }
+
+        return result.OfferDescriptionDatas;
     }
 }
