+++ src/marketplace/Apps.Service/BusinessLogic/IAppChangeBusinessLogic.cs
@@ -19,6 +19,7 @@
  ********************************************************************************/
 
 using Org.Eclipse.TractusX.Portal.Backend.Apps.Service.ViewModels;
+using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
 
 namespace Org.Eclipse.TractusX.Portal.Backend.Apps.Service.BusinessLogic;
 
@@ -27,12 +28,27 @@ namespace Org.Eclipse.TractusX.Portal.Backend.Apps.Service.BusinessLogic;
 /// </summary>
 public interface IAppChangeBusinessLogic
 {
-      /// <summary>
-      /// Add User Role for Active App and creates a notification
-      /// </summary>
-      /// <param name="appId"></param>
-      /// <param name="appUserRolesDescription"></param>
-      /// <param name="iamUserId"></param>
-      /// <returns>List of the created AppRoles</returns>
-      Task<IEnumerable<AppRoleData>> AddActiveAppUserRoleAsync(Guid appId, IEnumerable<AppUserRole> appUserRolesDescription, string iamUserId);
+    /// <summary>
+    /// Add User Role for Active App and creates a notification
+    /// </summary>
+    /// <param name="appId"></param>
+    /// <param name="appUserRolesDescription"></param>
+    /// <param name="iamUserId"></param>
+    /// <returns>List of the created AppRoles</returns>
+    Task<IEnumerable<AppRoleData>> AddActiveAppUserRoleAsync(Guid appId, IEnumerable<AppUserRole> appUserRolesDescription, string iamUserId);
+
+    /// <summary>
+    /// Get OfferDescription by appId
+    /// </summary>
+    /// <param name="appId">Id of the app</param>
+    /// <param name="iamUserId">Id of the iamUser</param>
+    Task<IEnumerable<LocalizedDescription>> GetAppUpdateDescriptionByIdAsync(Guid appId, string iamUserId);
+
+    /// <summary>
+    /// Create or Update OfferDescription by appId
+    /// </summary>
+    /// <param name="appId">Id of the app</param>
+    /// <param name="iamUserId">Id of the iamUser</param>
+    /// <param name="offerDescriptionDatas">OfferDescription Data</param>
+    Task CreateOrUpdateAppDescriptionByIdAsync(Guid appId, string iamUserId, IEnumerable<LocalizedDescription> offerDescriptionDatas);
 }
