+++ src/marketplace/Apps.Service/BusinessLogic/IAppsBusinessLogic.cs
@@ -156,21 +156,6 @@ public interface IAppsBusinessLogic
     Task<(byte[] Content, string ContentType, string FileName)> GetAppDocumentContentAsync(Guid appId, Guid documentId, CancellationToken cancellationToken);
 
     /// <summary>
-    /// Get OfferDescription by appId
-    /// </summary>
-    /// <param name="appId">Id of the app</param>
-    /// <param name="iamUserId">Id of the iamUser</param>
-    Task<IEnumerable<LocalizedDescription>> GetAppUpdateDescriptionByIdAsync(Guid appId, string iamUserId);
-
-    /// <summary>
-    /// Create or Update OfferDescription by appId
-    /// </summary>
-    /// <param name="appId">Id of the app</param>
-    /// <param name="iamUserId">Id of the iamUser</param>
-    /// <param name="offerDescriptionDatas">OfferDescription Data</param>
-    Task CreateOrUpdateAppDescriptionByIdAsync(Guid appId, string iamUserId, IEnumerable<LocalizedDescription> offerDescriptionDatas);
-
-    /// <summary>
     /// Create OfferAssigned AppLeadImage Document by appId
     /// </summary>
     /// <param name="appId">Id of the app</param>
